+++ generators/app/index.js
@@ -3,15 +3,15 @@
 // For Prompt options:
 // https://github.com/SBoudrias/Inquirer.js
 
-"use strict";
-const Generator = require("yeoman-generator");
-const chalk = require("chalk");
-const yosay = require("onionsay");
-const parseDomain = require("parse-domain");
-const niceware = require("niceware");
-const fsN = require("fs");
-const { v4: uuidv4 } = require("uuid");
-const bcrypt = require("bcrypt");
+'use strict';
+const Generator = require('yeoman-generator');
+const chalk = require('chalk');
+const yosay = require('onionsay');
+const { parseDomain } = require('parse-domain');
+const niceware = require('niceware');
+const fsN = require('fs');
+const { v4: uuidv4 } = require('uuid');
+const bcrypt = require('bcrypt');
 
 let defaultStore = false;
 let logger;
@@ -33,32 +33,32 @@ const isCorrectHostname = (hostname) => {
   var isCorrect = hostnameRegexp.test(hostname);
   return isCorrect
     ? true
-    : "Invalid hostname: should be something like host.example.org, myvm1, or somehost.amazonaws.com";
+    : 'Invalid hostname: should be something like host.example.org, myvm1, or somehost.amazonaws.com';
 };
-const em = (text) => chalk.keyword("orange")(text);
+const em = (text) => chalk.keyword('orange')(text);
 
 const defUseSubdomainPrompt = (a, service) => {
   var desc =
     servicesRolsMap[service].desc.length > 0
       ? ` (${servicesRolsMap[service].desc})`
-      : "";
+      : '';
   return `Will the ${em(service)} module${desc} use a http${
-    a.LA_enable_ssl ? "s" : ""
-  }://${em("subdomain")}.${a.LA_domain} or not?`;
+    a.LA_enable_ssl ? 's' : ''
+  }://${em('subdomain')}.${a.LA_domain} or not?`;
 };
 
 const validateDomain = (input, name) =>
   new Promise((resolve) => {
     if (debug) logger(`Validate ${input} ${name}`);
     // It's a domain not a https://url
-    const isValid = isCorrectDomain(input) && input.split("/").length === 1;
+    const isValid = isCorrectDomain(input) && input.split('/').length === 1;
     // As is a url we don't store now
     // if (isValid && store) storeMachine(name, input);
-    if (isValid || input === "other") {
+    if (isValid || input === 'other') {
       resolve(true);
     } else {
       if (debug) logger(input);
-      resolve("You need to provide some-example-domain.org");
+      resolve('You need to provide some-example-domain.org');
     }
   });
 
@@ -85,125 +85,125 @@ const vhostsSet = new Set(); // use for nginx_vhost_fragments_to_clear
 
 const servicesRolsMap = {
   main: {
-    name: "main",
-    group: "ala-demo",
-    playbook: "ala-demo-basic",
-    desc: "",
+    name: 'main',
+    group: 'ala-demo',
+    playbook: 'ala-demo-basic',
+    desc: '',
   },
   collectory: {
-    name: "collectory",
-    group: "collectory",
-    playbook: "collectory-standalone",
-    desc: "biodiversity collections",
+    name: 'collectory',
+    group: 'collectory',
+    playbook: 'collectory-standalone',
+    desc: 'biodiversity collections',
   },
   ala_hub: {
-    name: "ala_hub",
-    group: "biocache-hub",
-    playbook: "biocache-hub-standalone",
-    desc: "occurrences search frontend",
+    name: 'ala_hub',
+    group: 'biocache-hub',
+    playbook: 'biocache-hub-standalone',
+    desc: 'occurrences search frontend',
   },
   biocache_service: {
-    name: "biocache_service",
-    group: "biocache-service-clusterdb",
-    playbook: "biocache-service-clusterdb",
-    desc: "occurrences web service",
+    name: 'biocache_service',
+    group: 'biocache-service-clusterdb',
+    playbook: 'biocache-service-clusterdb',
+    desc: 'occurrences web service',
   },
   ala_bie: {
-    name: "ala_bie",
-    group: "bie-hub",
-    playbook: "bie-hub",
-    desc: "species search frontend",
+    name: 'ala_bie',
+    group: 'bie-hub',
+    playbook: 'bie-hub',
+    desc: 'species search frontend',
   },
   bie_index: {
-    name: "bie_index",
-    group: "bie-index",
-    playbook: "bie-index",
-    desc: "species web service",
+    name: 'bie_index',
+    group: 'bie-index',
+    playbook: 'bie-index',
+    desc: 'species web service',
   },
   images: {
-    name: "images",
-    group: "image-service",
-    playbook: "image-service",
-    desc: "",
+    name: 'images',
+    group: 'image-service',
+    playbook: 'image-service',
+    desc: '',
   },
   lists: {
-    name: "lists",
-    group: "species-list",
-    playbook: "species-list-standalone",
-    desc: "",
+    name: 'lists',
+    group: 'species-list',
+    playbook: 'species-list-standalone',
+    desc: '',
   },
   regions: {
-    name: "regions",
-    group: "regions",
-    playbook: "regions-standalone",
-    desc: "regional data frontend",
+    name: 'regions',
+    group: 'regions',
+    playbook: 'regions-standalone',
+    desc: 'regional data frontend',
   },
   logger: {
-    name: "logger",
-    group: "logger-service",
-    playbook: "logger-standalone",
-    desc: "event logging",
+    name: 'logger',
+    group: 'logger-service',
+    playbook: 'logger-standalone',
+    desc: 'event logging',
   },
   solr: {
-    name: "solr",
-    group: "solr7-server",
-    playbook: "solr7-standalone",
-    desc: "indexing",
+    name: 'solr',
+    group: 'solr7-server',
+    playbook: 'solr7-standalone',
+    desc: 'indexing',
   },
   cas: {
-    name: "cas",
-    group: "cas-servers",
-    playbook: "cas5-standalone",
-    desc: "authentication system",
+    name: 'cas',
+    group: 'cas-servers',
+    playbook: 'cas5-standalone',
+    desc: 'authentication system',
   },
   biocache_backend: {
-    name: "biocache_backend",
-    group: "biocache-db",
-    playbook: "biocache-db",
-    desc: "cassandra",
+    name: 'biocache_backend',
+    group: 'biocache-db',
+    playbook: 'biocache-db',
+    desc: 'cassandra',
   },
   biocache_cli: {
-    name: "biocache_cli",
-    group: "biocache-cli",
-    playbook: "biocache-cli",
+    name: 'biocache_cli',
+    group: 'biocache-cli',
+    playbook: 'biocache-cli',
     desc:
-      "manages the loading, sampling, processing and indexing of occurrence records",
+      'manages the loading, sampling, processing and indexing of occurrence records',
   },
   spatial: {
-    name: "spatial",
-    group: "spatial",
-    playbook: "spatial",
-    desc: "spatial front-end",
+    name: 'spatial',
+    group: 'spatial',
+    playbook: 'spatial',
+    desc: 'spatial front-end',
   },
   webapi: {
-    name: "webapi",
-    group: "webapi_standalone",
-    playbook: "webapi_standalone",
-    desc: "API front-end",
+    name: 'webapi',
+    group: 'webapi_standalone',
+    playbook: 'webapi_standalone',
+    desc: 'API front-end',
   },
   dashboard: {
-    name: "dashboard",
-    group: "dashboard",
-    playbook: "dashboard",
-    desc: "dashboard",
+    name: 'dashboard',
+    group: 'dashboard',
+    playbook: 'dashboard',
+    desc: 'dashboard',
   },
   alerts: {
-    name: "alerts",
-    group: "alerts-service",
-    playbook: "alerts-standalone",
-    desc: "alerts",
+    name: 'alerts',
+    group: 'alerts-service',
+    playbook: 'alerts-standalone',
+    desc: 'alerts',
   },
   doi: {
-    name: "doi",
-    group: "doi-service",
-    playbook: "doi-service-standalone",
-    desc: "DOI service",
+    name: 'doi',
+    group: 'doi-service',
+    playbook: 'doi-service-standalone',
+    desc: 'DOI service',
   },
   nameindexer: {
-    name: "nameindexer",
-    group: "nameindexer",
-    playbook: "nameindexer-standalone",
-    desc: "nameindexer",
+    name: 'nameindexer',
+    group: 'nameindexer',
+    playbook: 'nameindexer-standalone',
+    desc: 'nameindexer',
   },
 };
 
@@ -211,10 +211,10 @@ const storeMachine = (name, machine) =>
   new Promise((resolve, reject) => {
     if (debug) logger(`Store: ${name} -> ${machine}`);
     if (isCorrectHostname(machine) === true) {
-      if (debug) logger("We add to the first position");
+      if (debug) logger('We add to the first position');
       if (!machines.includes(machine)) machines.push(machine);
       if (!machinesSorted.includes(machine)) machinesSorted.unshift(machine);
-      if (name !== "main") {
+      if (name !== 'main') {
         servicesAndMachines.push({
           service: name,
           machine,
@@ -231,7 +231,7 @@ const storeMachine = (name, machine) =>
 
 function PromptSubdomainFor(name, subdomain, when, force) {
   this.store = defaultStore;
-  this.type = "confirm";
+  this.type = 'confirm';
   const varName = `LA_${name}_uses_subdomain`;
   this.name = varName;
   this.message = (a) => defUseSubdomainPrompt(a, name);
@@ -251,12 +251,12 @@ function PromptSubdomainFor(name, subdomain, when, force) {
 
 function urlDefValue(a, varUsesSubdomain, subdomain, varName) {
   const hasPrevious =
-    typeof previousConfig !== "undefined" &&
-    typeof previousConfig[varName] !== "undefined";
+    typeof previousConfig !== 'undefined' &&
+    typeof previousConfig[varName] !== 'undefined';
   if (replay && hasPrevious) {
     return previousConfig[varName];
   }
-  if (typeof varUsesSubdomain !== "undefined" && a[varUsesSubdomain]) {
+  if (typeof varUsesSubdomain !== 'undefined' && a[varUsesSubdomain]) {
     return `${subdomain}.${a.LA_domain}`;
   }
   return `${a.LA_domain}`;
@@ -264,15 +264,15 @@ function urlDefValue(a, varUsesSubdomain, subdomain, varName) {
 
 function hostChoices(a, varUsesSubdomain, subdomain, varName) {
   let choices;
-  if (typeof varUsesSubdomain !== "undefined" && a[varUsesSubdomain]) {
-    choices = [...machinesSorted, `${subdomain}.${a.LA_domain}`, "other"];
+  if (typeof varUsesSubdomain !== 'undefined' && a[varUsesSubdomain]) {
+    choices = [...machinesSorted, `${subdomain}.${a.LA_domain}`, 'other'];
   } else {
-    choices = [...machinesSorted, "other"];
+    choices = [...machinesSorted, 'other'];
   }
   // Add previous selected hostname at the start of the array
   const hasPrevious =
-    typeof previousConfig !== "undefined" &&
-    typeof previousConfig[varName] !== "undefined";
+    typeof previousConfig !== 'undefined' &&
+    typeof previousConfig[varName] !== 'undefined';
   const previousHostname = hasPrevious ? previousConfig[varName] : null;
   if (debug) logger(`hasPrevious: ${hasPrevious} ${previousHostname}`);
   if (replay && hasPrevious) {
@@ -292,7 +292,7 @@ function hostChoices(a, varUsesSubdomain, subdomain, varName) {
 
 function PromptHostnameFor(name, subdomain, when) {
   this.store = defaultStore;
-  this.type = "list";
+  this.type = 'list';
   const varName = `LA_${name}_hostname`;
   const varUsesSubdomain = `LA_${name}_uses_subdomain`;
   this.name = varName;
@@ -310,17 +310,17 @@ function PromptHostnameFor(name, subdomain, when) {
 
 function PromptHostnameInputFor(name, when) {
   this.store = defaultStore;
-  this.type = "input";
+  this.type = 'input';
   const varName = `LA_${name}_hostname`;
   this.name = varName;
   this.message = `LA ${em(name)} hostname ?`;
   this.when = (a) => {
     if (debug) logger(`We ask for other ${name}/${a[varName]}?`);
-    if (typeof when !== "undefined" && !when(a)) {
+    if (typeof when !== 'undefined' && !when(a)) {
       if (debug) logger("We don't ask for other hostname");
       return false;
     }
-    if (a[varName] === "other") {
+    if (a[varName] === 'other') {
       if (debug) logger("We ask for 'other' hostname");
       return true;
     }
@@ -333,7 +333,7 @@ function PromptHostnameInputFor(name, when) {
 
 function PromptUrlFor(name, subdomain, when) {
   this.store = defaultStore;
-  this.type = "input";
+  this.type = 'input';
   const varName = `LA_${name}_url`;
   const varUsesSubdomain = `LA_${name}_uses_subdomain`;
   this.name = varName;
@@ -360,14 +360,14 @@ function PromptPathFor(name, path, when) {
   const varUsesSubdomain = `LA_${name}_uses_subdomain`;
   const varHostname = `LA_${name}_hostname`;
   const varUrl = `LA_${name}_url`;
-  this.type = "input";
+  this.type = 'input';
   this.name = varName;
   this.message = (a) => {
     // Warn: dup code below
     const samplePath = a[varUsesSubdomain]
-      ? "/"
-      : `/${typeof path === "undefined" ? "" : path}`;
-    return `Which context ${em("/path")} you wanna use for this service (like ${
+      ? '/'
+      : `/${typeof path === 'undefined' ? '' : path}`;
+    return `Which context ${em('/path')} you wanna use for this service (like ${
       a[varUrl]
     }${em(samplePath)}) ?`;
   };
@@ -376,12 +376,12 @@ function PromptPathFor(name, path, when) {
       const hostname = a[varHostname];
       const url = a[varUrl];
       const rootUsed =
-        typeof machinesAndPaths[hostname] !== "undefined" &&
-        typeof machinesAndPaths[hostname][url] !== "undefined" &&
-        machinesAndPaths[hostname][url]["/"];
-      return rootUsed ? `/${typeof path === "undefined" ? "" : path}` : "/";
+        typeof machinesAndPaths[hostname] !== 'undefined' &&
+        typeof machinesAndPaths[hostname][url] !== 'undefined' &&
+        machinesAndPaths[hostname][url]['/'];
+      return rootUsed ? `/${typeof path === 'undefined' ? '' : path}` : '/';
     }
-    return typeof path === "undefined" ? "" : `/${path}`;
+    return typeof path === 'undefined' ? '' : `/${path}`;
   };
   if (when) {
     this.when = when;
@@ -389,7 +389,7 @@ function PromptPathFor(name, path, when) {
   this.validate = (path, answers) =>
     new Promise((resolve) => {
       // Check if in this machine this path is already used
-      if (!path.startsWith("/")) {
+      if (!path.startsWith('/')) {
         resolve("Please enter something like '/path'");
       }
       // FIXME check better url here
@@ -417,49 +417,49 @@ module.exports = class extends Generator {
   constructor(args, opts) {
     super(args, opts);
 
-    this.argument("debug", { type: Boolean, required: false });
-    this.argument("replay", { type: Boolean, required: false });
-    this.argument("replay-dont-ask", { type: Boolean, required: false });
+    this.argument('debug', { type: Boolean, required: false });
+    this.argument('replay', { type: Boolean, required: false });
+    this.argument('replay-dont-ask', { type: Boolean, required: false });
 
     debug = this.options.debug;
-    replay = this.options.replay || this.options["replay-dont-ask"];
-    dontAsk = this.options["replay-dont-ask"];
+    replay = this.options.replay || this.options['replay-dont-ask'];
+    dontAsk = this.options['replay-dont-ask'];
 
     const previousConfigAll = this.config.getAll();
     previousConfig =
-      typeof previousConfigAll === "undefined"
+      typeof previousConfigAll === 'undefined'
         ? []
         : previousConfigAll.promptValues;
 
     firstRun = previousConfigAll.firstRun !== false;
     if (firstRun) {
       // Set firstRun so in the future we can check it
-      this.config.set("firstRun", false);
+      this.config.set('firstRun', false);
     }
 
     logger = this.log;
 
-    let cmdResult = this.spawnCommandSync("which", ["git"], {
+    let cmdResult = this.spawnCommandSync('which', ['git'], {
       shell: true,
       stdio: null,
     });
     if (cmdResult.status !== 0) {
       logger(
-        `${em("Error")}: Please install git before running this generator`
+        `${em('Error')}: Please install git before running this generator`
       );
       process.exit(1);
     }
 
     // We always store in the first run
     if (debug && firstRun) {
-      logger("Running generator for the first time in this location");
+      logger('Running generator for the first time in this location');
     }
     defaultStore = firstRun || this.options.replay;
-    if (debug) logger("Current config:");
+    if (debug) logger('Current config:');
     if (debug) logger(previousConfig);
-    if (typeof previousConfig === "undefined" && replay) {
+    if (typeof previousConfig === 'undefined' && replay) {
       logger(
-        `${em("Error")}: We cannot replay without a previous configuration`
+        `${em('Error')}: We cannot replay without a previous configuration`
       );
       process.exit(1);
     }
@@ -469,7 +469,7 @@ module.exports = class extends Generator {
     // Have Yeoman greet the user.
     logger(
       yosay(
-        `Welcome to the ${em("Living Atlas")} Quick-Start Ansible Generator`
+        `Welcome to the ${em('Living Atlas')} Quick-Start Ansible Generator`
       )
     );
 
@@ -478,165 +478,165 @@ module.exports = class extends Generator {
       : await this.prompt([
           {
             store: true,
-            type: "input",
-            name: "LA_project_name",
-            message: `Your LA Project ${em("Long Name")}:`,
-            default: "Living Atlas of Wakanda",
+            type: 'input',
+            name: 'LA_project_name',
+            message: `Your LA Project ${em('Long Name')}:`,
+            default: 'Living Atlas of Wakanda',
           },
           {
             store: true,
-            type: "input",
-            name: "LA_project_shortname",
-            message: `Your LA Project ${em("Shortname")}:`,
+            type: 'input',
+            name: 'LA_project_shortname',
+            message: `Your LA Project ${em('Shortname')}:`,
             default: (answers) =>
-              answers.LA_project_name.replace(/Living Atlas of /g, "LA "),
+              answers.LA_project_name.replace(/Living Atlas of /g, 'LA '),
           },
           {
             store: true,
-            type: "input",
-            name: "LA_pkg_name",
+            type: 'input',
+            name: 'LA_pkg_name',
             message: `Your LA ${em(
-              "short-lowercase-name"
+              'short-lowercase-name'
             )} (we'll put your inventories in that directory):`,
             default: (answers) =>
-              answers.LA_project_name.toLowerCase().replace(/ /g, "-"),
+              answers.LA_project_name.toLowerCase().replace(/ /g, '-'),
             validate: (input) =>
               new Promise((resolve) => {
                 if (input.match(/^[a-z0-9-]+$/g)) {
                   resolve(true);
                 } else {
-                  resolve("You need to provide some-example-short-name");
+                  resolve('You need to provide some-example-short-name');
                 }
               }),
           },
           {
             store: true,
-            type: "input",
-            name: "LA_domain",
-            message: `What is your LA node ${em("main domain")}?`,
+            type: 'input',
+            name: 'LA_domain',
+            message: `What is your LA node ${em('main domain')}?`,
             default: (answers) => `${answers.LA_pkg_name}.org`,
-            validate: (input) => validateDomain(input, "main", true),
+            validate: (input) => validateDomain(input, 'main', true),
           },
           {
             store: true,
-            type: "confirm",
-            name: "LA_use_species",
-            message: `Use ${em("species")} service?`,
+            type: 'confirm',
+            name: 'LA_use_species',
+            message: `Use ${em('species')} service?`,
             default: true,
           },
           {
             store: true,
-            type: "confirm",
-            name: "LA_use_species_lists",
-            message: `Use ${em("specieslists")} service?`,
+            type: 'confirm',
+            name: 'LA_use_species_lists',
+            message: `Use ${em('specieslists')} service?`,
             when: (a) => a.LA_use_species,
             default: true,
           },
           {
             store: true,
-            type: "confirm",
-            name: "LA_use_images",
-            message: `Use ${em("images")} service?`,
+            type: 'confirm',
+            name: 'LA_use_images',
+            message: `Use ${em('images')} service?`,
             default: true,
           },
           {
             store: true,
-            type: "confirm",
-            name: "LA_use_spatial",
-            message: `Use ${em("spatial")} service?`,
+            type: 'confirm',
+            name: 'LA_use_spatial',
+            message: `Use ${em('spatial')} service?`,
             default: true,
           },
           {
             store: true,
-            type: "confirm",
-            name: "LA_use_regions",
-            message: `Use ${em("regions")} service?`,
+            type: 'confirm',
+            name: 'LA_use_regions',
+            message: `Use ${em('regions')} service?`,
             default: true,
           },
 
           {
             store: true,
-            type: "confirm",
-            name: "LA_use_webapi",
+            type: 'confirm',
+            name: 'LA_use_webapi',
             message: `Use ${em(
-              "webapi"
+              'webapi'
             )} an API documentation service (similar to api.ala.org.au but empty)?`,
             default: false,
           },
           {
             store: true,
-            type: "confirm",
-            name: "LA_use_dashboard",
+            type: 'confirm',
+            name: 'LA_use_dashboard',
             message: `Use ${em(
-              "dashboard"
+              'dashboard'
             )} view stats service (similar to dashboard.ala.org.au)?`,
             default: false,
           },
           {
             store: true,
-            type: "confirm",
-            name: "LA_use_alerts",
-            message: `Use ${em("alerts")} service?`,
+            type: 'confirm',
+            name: 'LA_use_alerts',
+            message: `Use ${em('alerts')} service?`,
             default: false,
           },
           {
             store: true,
-            type: "confirm",
-            name: "LA_use_doi",
-            message: `Use ${em("doi")} service?`,
+            type: 'confirm',
+            name: 'LA_use_doi',
+            message: `Use ${em('doi')} service?`,
             default: false,
           },
           {
             store: true,
-            type: "confirm",
-            name: "LA_use_CAS",
-            message: `Use ${em("CAS")} Auth service?`,
+            type: 'confirm',
+            name: 'LA_use_CAS',
+            message: `Use ${em('CAS')} Auth service?`,
             default: true,
           },
           {
             store: true,
-            type: "confirm",
-            name: "LA_enable_ssl",
-            message: `Enable ${em("SSL")}?`,
+            type: 'confirm',
+            name: 'LA_enable_ssl',
+            message: `Enable ${em('SSL')}?`,
             default: true,
           },
           {
             store: defaultStore,
-            type: "input",
-            name: "check-ssl",
-            message: "",
-            default: "",
+            type: 'input',
+            name: 'check-ssl',
+            message: '',
+            default: '',
             when: (a) =>
               new Promise((resolve) => {
-                a.LA_urls_prefix = a.LA_enable_ssl ? "https://" : "http://";
+                a.LA_urls_prefix = a.LA_enable_ssl ? 'https://' : 'http://';
                 resolve(false);
               }),
           },
           {
             store: true,
-            type: "input",
-            name: "LA_main_hostname",
+            type: 'input',
+            name: 'LA_main_hostname',
             message: `Hostname for the basic LA branding?`,
-            filter: (input) => storeMachine("main", input),
+            filter: (input) => storeMachine('main', input),
             validate: (input) => isCorrectHostname(input),
             default: (a) => a.LA_domain,
           },
 
-          new PromptSubdomainFor("cas", "auth", true, true),
-          new PromptHostnameFor("cas", "auth"),
-          new PromptHostnameInputFor("cas"),
-          new PromptUrlFor("cas", "auth"),
+          new PromptSubdomainFor('cas', 'auth', true, true),
+          new PromptHostnameFor('cas', 'auth'),
+          new PromptHostnameInputFor('cas'),
+          new PromptUrlFor('cas', 'auth'),
           {
             store: true,
-            type: "input",
-            name: "LA_biocache_backend_hostname",
-            message: `LA ${em("biocache-backend")} hostname`,
+            type: 'input',
+            name: 'LA_biocache_backend_hostname',
+            message: `LA ${em('biocache-backend')} hostname`,
             validate: (input) => isCorrectHostname(input),
             filter: (input) =>
               new Promise((resolve) => {
-                storeMachine("biocache_backend", input).then((input) =>
-                  storeMachine("biocache_cli", input).then((input) =>
-                    storeMachine("nameindexer", input).then((input) =>
+                storeMachine('biocache_backend', input).then((input) =>
+                  storeMachine('biocache_cli', input).then((input) =>
+                    storeMachine('nameindexer', input).then((input) =>
                       resolve(input)
                     )
                   )
@@ -646,218 +646,218 @@ module.exports = class extends Generator {
           },
           {
             store: defaultStore,
-            type: "list",
-            name: "LA_collectory_uses_subdomain",
+            type: 'list',
+            name: 'LA_collectory_uses_subdomain',
             message: (a) =>
-              `Will the ${em("collectory")} service use a ${
+              `Will the ${em('collectory')} service use a ${
                 a.LA_urls_prefix
-              }${em("subdomain")}.${a.LA_domain} or a ${a.LA_urls_prefix}${
+              }${em('subdomain')}.${a.LA_domain} or a ${a.LA_urls_prefix}${
                 a.LA_domain
-              }${em("/service-path")} ?`,
+              }${em('/service-path')} ?`,
             choices: [
-              { name: "subdomain", value: true },
-              { name: "service-path", value: false },
+              { name: 'subdomain', value: true },
+              { name: 'service-path', value: false },
             ],
           },
-          new PromptHostnameFor("collectory", "collections"),
-          new PromptHostnameInputFor("collectory"),
-          new PromptUrlFor("collectory", "collections"),
-          new PromptPathFor("collectory", "collections"),
+          new PromptHostnameFor('collectory', 'collections'),
+          new PromptHostnameInputFor('collectory'),
+          new PromptUrlFor('collectory', 'collections'),
+          new PromptPathFor('collectory', 'collections'),
 
-          new PromptSubdomainFor("ala_hub", "records"),
-          new PromptHostnameFor("ala_hub", "records"),
-          new PromptHostnameInputFor("ala_hub"),
-          new PromptUrlFor("ala_hub", "records"),
-          new PromptPathFor("ala_hub", "records"),
+          new PromptSubdomainFor('ala_hub', 'records'),
+          new PromptHostnameFor('ala_hub', 'records'),
+          new PromptHostnameInputFor('ala_hub'),
+          new PromptUrlFor('ala_hub', 'records'),
+          new PromptPathFor('ala_hub', 'records'),
 
-          new PromptSubdomainFor("biocache_service", "records-ws"),
-          new PromptHostnameFor("biocache_service", "records-ws"),
-          new PromptHostnameInputFor("biocache_service"),
-          new PromptUrlFor("biocache_service", "records-ws"),
-          new PromptPathFor("biocache_service", "records-ws"),
+          new PromptSubdomainFor('biocache_service', 'records-ws'),
+          new PromptHostnameFor('biocache_service', 'records-ws'),
+          new PromptHostnameInputFor('biocache_service'),
+          new PromptUrlFor('biocache_service', 'records-ws'),
+          new PromptPathFor('biocache_service', 'records-ws'),
 
-          new PromptSubdomainFor("ala_bie", "species", (a) => a.LA_use_species),
-          new PromptHostnameFor("ala_bie", "species", (a) => a.LA_use_species),
-          new PromptHostnameInputFor("ala_bie", (a) => a.LA_use_species),
-          new PromptUrlFor("ala_bie", "species", (a) => a.LA_use_species),
-          new PromptPathFor("ala_bie", "species", (a) => a.LA_use_species),
+          new PromptSubdomainFor('ala_bie', 'species', (a) => a.LA_use_species),
+          new PromptHostnameFor('ala_bie', 'species', (a) => a.LA_use_species),
+          new PromptHostnameInputFor('ala_bie', (a) => a.LA_use_species),
+          new PromptUrlFor('ala_bie', 'species', (a) => a.LA_use_species),
+          new PromptPathFor('ala_bie', 'species', (a) => a.LA_use_species),
 
           new PromptSubdomainFor(
-            "bie_index",
-            "species-service",
+            'bie_index',
+            'species-service',
             (a) => a.LA_use_species
           ),
           new PromptHostnameFor(
-            "bie_index",
-            "species-ws",
+            'bie_index',
+            'species-ws',
             (a) => a.LA_use_species
           ),
-          new PromptHostnameInputFor("bie_index", (a) => a.LA_use_species),
-          new PromptUrlFor("bie_index", "species-ws", (a) => a.LA_use_species),
-          new PromptPathFor("bie_index", "species-ws", (a) => a.LA_use_species),
+          new PromptHostnameInputFor('bie_index', (a) => a.LA_use_species),
+          new PromptUrlFor('bie_index', 'species-ws', (a) => a.LA_use_species),
+          new PromptPathFor('bie_index', 'species-ws', (a) => a.LA_use_species),
 
-          new PromptSubdomainFor("images", "images", (a) => a.LA_use_images),
-          new PromptHostnameFor("images", "images", (a) => a.LA_use_images),
-          new PromptHostnameInputFor("images", (a) => a.LA_use_images),
-          new PromptUrlFor("images", "images", (a) => a.LA_use_images),
-          new PromptPathFor("images", "images", (a) => a.LA_use_images),
+          new PromptSubdomainFor('images', 'images', (a) => a.LA_use_images),
+          new PromptHostnameFor('images', 'images', (a) => a.LA_use_images),
+          new PromptHostnameInputFor('images', (a) => a.LA_use_images),
+          new PromptUrlFor('images', 'images', (a) => a.LA_use_images),
+          new PromptPathFor('images', 'images', (a) => a.LA_use_images),
 
           new PromptSubdomainFor(
-            "lists",
-            "specieslists",
+            'lists',
+            'specieslists',
             (a) => a.LA_use_species_lists
           ),
           new PromptHostnameFor(
-            "lists",
-            "lists",
+            'lists',
+            'lists',
             (a) => a.LA_use_species_lists
           ),
-          new PromptHostnameInputFor("lists", (a) => a.LA_use_species_lists),
-          new PromptUrlFor("lists", "lists", (a) => a.LA_use_species_lists),
+          new PromptHostnameInputFor('lists', (a) => a.LA_use_species_lists),
+          new PromptUrlFor('lists', 'lists', (a) => a.LA_use_species_lists),
           new PromptPathFor(
-            "lists",
-            "specieslists",
+            'lists',
+            'specieslists',
             (a) => a.LA_use_species_lists
           ),
 
-          new PromptSubdomainFor("regions", "regions", (a) => a.LA_use_regions),
-          new PromptHostnameFor("regions", "regions", (a) => a.LA_use_regions),
-          new PromptHostnameInputFor("regions", (a) => a.LA_use_regions),
-          new PromptUrlFor("regions", "regions", (a) => a.LA_use_regions),
-          new PromptPathFor("regions", "regions", (a) => a.LA_use_regions),
+          new PromptSubdomainFor('regions', 'regions', (a) => a.LA_use_regions),
+          new PromptHostnameFor('regions', 'regions', (a) => a.LA_use_regions),
+          new PromptHostnameInputFor('regions', (a) => a.LA_use_regions),
+          new PromptUrlFor('regions', 'regions', (a) => a.LA_use_regions),
+          new PromptPathFor('regions', 'regions', (a) => a.LA_use_regions),
 
-          new PromptSubdomainFor("logger", "logger"),
-          new PromptHostnameFor("logger", "logger"),
-          new PromptHostnameInputFor("logger"),
-          new PromptUrlFor("logger", "logger"),
-          new PromptPathFor("logger", "logger-service"),
+          new PromptSubdomainFor('logger', 'logger'),
+          new PromptHostnameFor('logger', 'logger'),
+          new PromptHostnameInputFor('logger'),
+          new PromptUrlFor('logger', 'logger'),
+          new PromptPathFor('logger', 'logger-service'),
 
-          new PromptSubdomainFor("webapi", "webapi", (a) => a.LA_use_webapi),
-          new PromptHostnameFor("webapi", "api", (a) => a.LA_use_webapi),
-          new PromptHostnameInputFor("webapi", (a) => a.LA_use_webapi),
-          new PromptUrlFor("webapi", "api", (a) => a.LA_use_webapi),
-          new PromptPathFor("webapi", "webapi", (a) => a.LA_use_webapi),
+          new PromptSubdomainFor('webapi', 'webapi', (a) => a.LA_use_webapi),
+          new PromptHostnameFor('webapi', 'api', (a) => a.LA_use_webapi),
+          new PromptHostnameInputFor('webapi', (a) => a.LA_use_webapi),
+          new PromptUrlFor('webapi', 'api', (a) => a.LA_use_webapi),
+          new PromptPathFor('webapi', 'webapi', (a) => a.LA_use_webapi),
 
           new PromptSubdomainFor(
-            "dashboard",
-            "dashboard",
+            'dashboard',
+            'dashboard',
             (a) => a.LA_use_dashboard
           ),
           new PromptHostnameFor(
-            "dashboard",
-            "dashboard",
+            'dashboard',
+            'dashboard',
             (a) => a.LA_use_dashboard
           ),
-          new PromptHostnameInputFor("dashboard", (a) => a.LA_use_dashboard),
-          new PromptUrlFor("dashboard", "dashboard", (a) => a.LA_use_dashboard),
+          new PromptHostnameInputFor('dashboard', (a) => a.LA_use_dashboard),
+          new PromptUrlFor('dashboard', 'dashboard', (a) => a.LA_use_dashboard),
           new PromptPathFor(
-            "dashboard",
-            "dashboard",
+            'dashboard',
+            'dashboard',
             (a) => a.LA_use_dashboard
           ),
 
-          new PromptSubdomainFor("alerts", "alerts", (a) => a.LA_use_alerts),
-          new PromptHostnameFor("alerts", "alerts", (a) => a.LA_use_alerts),
-          new PromptHostnameInputFor("alerts", (a) => a.LA_use_alerts),
-          new PromptUrlFor("alerts", "alerts", (a) => a.LA_use_alerts),
-          new PromptPathFor("alerts", "alerts", (a) => a.LA_use_alerts),
+          new PromptSubdomainFor('alerts', 'alerts', (a) => a.LA_use_alerts),
+          new PromptHostnameFor('alerts', 'alerts', (a) => a.LA_use_alerts),
+          new PromptHostnameInputFor('alerts', (a) => a.LA_use_alerts),
+          new PromptUrlFor('alerts', 'alerts', (a) => a.LA_use_alerts),
+          new PromptPathFor('alerts', 'alerts', (a) => a.LA_use_alerts),
 
-          new PromptSubdomainFor("doi", "doi", (a) => a.LA_use_doi),
-          new PromptHostnameFor("doi", "doi", (a) => a.LA_use_doi),
-          new PromptHostnameInputFor("doi", (a) => a.LA_use_doi),
-          new PromptUrlFor("doi", "doi", (a) => a.LA_use_doi),
-          new PromptPathFor("doi", "doi", (a) => a.LA_use_doi),
+          new PromptSubdomainFor('doi', 'doi', (a) => a.LA_use_doi),
+          new PromptHostnameFor('doi', 'doi', (a) => a.LA_use_doi),
+          new PromptHostnameInputFor('doi', (a) => a.LA_use_doi),
+          new PromptUrlFor('doi', 'doi', (a) => a.LA_use_doi),
+          new PromptPathFor('doi', 'doi', (a) => a.LA_use_doi),
 
-          new PromptSubdomainFor("solr", "solr"),
-          new PromptHostnameFor("solr", "index"),
-          new PromptHostnameInputFor("solr"),
-          new PromptUrlFor("solr", "index"),
-          new PromptPathFor("solr", "solr"),
+          new PromptSubdomainFor('solr', 'solr'),
+          new PromptHostnameFor('solr', 'index'),
+          new PromptHostnameInputFor('solr'),
+          new PromptUrlFor('solr', 'index'),
+          new PromptPathFor('solr', 'solr'),
 
           new PromptSubdomainFor(
-            "spatial",
-            "spatial",
+            'spatial',
+            'spatial',
             (a) => a.LA_use_spatial,
             true
           ),
-          new PromptHostnameFor("spatial", "spatial", (a) => a.LA_use_spatial),
-          new PromptHostnameInputFor("spatial", (a) => a.LA_use_spatial),
-          new PromptUrlFor("spatial", "spatial", (a) => a.LA_use_spatial),
+          new PromptHostnameFor('spatial', 'spatial', (a) => a.LA_use_spatial),
+          new PromptHostnameInputFor('spatial', (a) => a.LA_use_spatial),
+          new PromptUrlFor('spatial', 'spatial', (a) => a.LA_use_spatial),
           {
             store: true,
-            type: "confirm",
-            name: "LA_generate_branding",
+            type: 'confirm',
+            name: 'LA_generate_branding',
             message: `Do you want to generate also a sample compatible ${em(
-              "LA branding"
+              'LA branding'
             )}? (Recommended to start, later you can improve it with your site style)`,
             default: true,
           },
           {
             store: true,
-            type: "confirm",
-            name: "LA_use_git",
+            type: 'confirm',
+            name: 'LA_use_git',
             message: `Use ${em(
-              "git"
+              'git'
             )} in your generated inventories to track their changes? (Very recommended)`,
             default: true,
           },
         ]);
 
     // For back-compatibility
-    if (typeof this.answers.LA_main_hostname === "undefined") {
+    if (typeof this.answers.LA_main_hostname === 'undefined') {
       this.answers.LA_main_hostname = this.answers.LA_domain;
     }
     this.answers.LA_biocache_cli_hostname = this.answers.LA_biocache_backend_hostname;
     this.answers.LA_nameindexer_hostname = this.answers.LA_biocache_backend_hostname;
 
-    if (typeof this.answers.LA_spatial_uses_subdomain === "undefined")
+    if (typeof this.answers.LA_spatial_uses_subdomain === 'undefined')
       this.answers.LA_spatial_uses_subdomain = true;
-    if (typeof this.answers.LA_cas_uses_subdomain === "undefined")
+    if (typeof this.answers.LA_cas_uses_subdomain === 'undefined')
       this.answers.LA_cas_uses_subdomain = true;
 
-    if (typeof this.answers.LA_use_images === "undefined")
+    if (typeof this.answers.LA_use_images === 'undefined')
       this.answers.LA_use_images = true;
-    if (typeof this.answers.LA_use_species_lists === "undefined")
+    if (typeof this.answers.LA_use_species_lists === 'undefined')
       this.answers.LA_use_species_lists = false;
-    if (typeof this.answers.LA_use_species === "undefined")
+    if (typeof this.answers.LA_use_species === 'undefined')
       this.answers.LA_use_species = true;
 
     if (dontAsk) {
       // Compatible with old generated inventories and don-ask
-      if (typeof this.answers.LA_use_webapi === "undefined")
+      if (typeof this.answers.LA_use_webapi === 'undefined')
         this.answers.LA_use_webapi = false;
 
-      if (typeof this.answers.LA_use_alerts === "undefined")
+      if (typeof this.answers.LA_use_alerts === 'undefined')
         this.answers.LA_use_alerts = false;
 
-      if (typeof this.answers.LA_use_doi === "undefined")
+      if (typeof this.answers.LA_use_doi === 'undefined')
         this.answers.LA_use_doi = false;
 
-      if (typeof this.answers.LA_use_dashboard === "undefined")
+      if (typeof this.answers.LA_use_dashboard === 'undefined')
         this.answers.LA_use_dashboard = false;
 
-      if (typeof this.answers.LA_generate_branding === "undefined")
+      if (typeof this.answers.LA_generate_branding === 'undefined')
         this.answers.LA_generate_branding = false;
 
       this.answers.LA_urls_prefix = this.answers.LA_enable_ssl
-        ? "https://"
-        : "http://";
+        ? 'https://'
+        : 'http://';
 
       vhostsSet.add(this.answers.LA_domain);
 
       if (debug) logger(this.answers);
 
       Object.keys(servicesRolsMap).forEach((service) => {
-        if (service === "spatial" && !this.answers.LA_use_spatial) return;
-        if (service === "regions" && !this.answers.LA_use_regions) return;
-        if (service === "ala_bie" && !this.answers.LA_use_species) return;
-        if (service === "bie_index" && !this.answers.LA_use_species) return;
-        if (service === "lists" && !this.answers.LA_use_species_lists) return;
-        if (service === "images" && !this.answers.LA_use_images) return;
-        if (service === "webapi" && !this.answers.LA_use_webapi) return;
-        if (service === "alerts" && !this.answers.LA_use_alerts) return;
-        if (service === "doi" && !this.answers.LA_use_doi) return;
-        if (service === "dashboard" && !this.answers.LA_use_dashboard) return;
+        if (service === 'spatial' && !this.answers.LA_use_spatial) return;
+        if (service === 'regions' && !this.answers.LA_use_regions) return;
+        if (service === 'ala_bie' && !this.answers.LA_use_species) return;
+        if (service === 'bie_index' && !this.answers.LA_use_species) return;
+        if (service === 'lists' && !this.answers.LA_use_species_lists) return;
+        if (service === 'images' && !this.answers.LA_use_images) return;
+        if (service === 'webapi' && !this.answers.LA_use_webapi) return;
+        if (service === 'alerts' && !this.answers.LA_use_alerts) return;
+        if (service === 'doi' && !this.answers.LA_use_doi) return;
+        if (service === 'dashboard' && !this.answers.LA_use_dashboard) return;
 
         const hostVar = `LA_${service}_hostname`;
         const serviceUrl = this.answers[`LA_${service}_url`];
@@ -865,7 +865,7 @@ module.exports = class extends Generator {
         if (debug)
           logger(`${service}: ${hostVar} -> ${hostname}, url: ${serviceUrl}`);
         storeMachine(service, hostname);
-        if (typeof serviceUrl !== "undefined") vhostsSet.add(serviceUrl);
+        if (typeof serviceUrl !== 'undefined') vhostsSet.add(serviceUrl);
       });
 
       if (debug) logger(machines);
@@ -881,57 +881,57 @@ module.exports = class extends Generator {
     const cmdOpts = {
       cwd: this.destinationPath(dest),
       shell: true,
-      stdio: "inherit",
+      stdio: 'inherit',
     };
 
     if (debug) logger(`Destination root: ${this.destinationRoot()}`);
     if (debug) logger(`cmdOpts: ${JSON.stringify(cmdOpts)}`);
 
     const services = [
-      "collectory",
-      "ala_hub",
-      "biocache_service",
-      "biocache_backend",
-      "biocache_cli",
-      "nameindexer",
-      "ala_bie",
-      "bie_index",
-      "images",
-      "logger",
-      "lists",
-      "regions",
-      "webapi",
-      "alerts",
-      "doi",
-      "dashboard",
+      'collectory',
+      'ala_hub',
+      'biocache_service',
+      'biocache_backend',
+      'biocache_cli',
+      'nameindexer',
+      'ala_bie',
+      'bie_index',
+      'images',
+      'logger',
+      'lists',
+      'regions',
+      'webapi',
+      'alerts',
+      'doi',
+      'dashboard',
     ];
 
     services.forEach((service) => {
       const path = this.answers[`LA_${service}_path`];
-      if (path === "/") {
-        this.answers[`LA_${service}_path`] = "";
+      if (path === '/') {
+        this.answers[`LA_${service}_path`] = '';
       }
       // url var is new, so we use hostname for old generated inventories
-      if (typeof this.answers[`LA_${service}_url`] === "undefined") {
+      if (typeof this.answers[`LA_${service}_url`] === 'undefined') {
         this.answers[`LA_${service}_url`] = this.answers[
           `LA_${service}_hostname`
         ];
       }
     });
     if (this.answers.LA_solr_uses_subdomain) {
-      this.answers.LA_solr_path = "";
+      this.answers.LA_solr_path = '';
     }
     // Backward compatibility
-    if (typeof this.answers.LA_solr_url === "undefined") {
+    if (typeof this.answers.LA_solr_url === 'undefined') {
       this.answers.LA_solr_url = this.answers.LA_solr_hostname;
     }
-    if (typeof this.answers.LA_cas_url === "undefined") {
+    if (typeof this.answers.LA_cas_url === 'undefined') {
       this.answers.LA_cas_url = this.answers.LA_cas_hostname;
     }
-    if (typeof this.answers.LA_spatial_url === "undefined") {
+    if (typeof this.answers.LA_spatial_url === 'undefined') {
       this.answers.LA_spatial_url = this.answers.LA_spatial_hostname;
     }
-    if (typeof this.answers.LA_generate_branding === "undefined") {
+    if (typeof this.answers.LA_generate_branding === 'undefined') {
       this.answers.LA_generate_branding = false;
     }
     this.answers.LA_machines = machines;
@@ -948,19 +948,19 @@ module.exports = class extends Generator {
     if (debug) logger(this.answers);
 
     this.fs.copyTpl(
-      this.templatePath("README-main.md"),
-      this.destinationPath("README.md"),
+      this.templatePath('README-main.md'),
+      this.destinationPath('README.md'),
       this.answers
     );
     this.fs.copyTpl(
-      this.templatePath("README.md"),
+      this.templatePath('README.md'),
       this.destinationPath(`${dest}/README.md`),
       this.answers
     );
 
     if (!this.fs.exists(`${dest}/dot-ssh-config`)) {
       this.fs.copyTpl(
-        this.templatePath("dot-ssh-config.sample"),
+        this.templatePath('dot-ssh-config.sample'),
         this.destinationPath(`${dest}/dot-ssh-config`),
         this.answers
       );
@@ -968,10 +968,10 @@ module.exports = class extends Generator {
 
     // Migrate old quick-start generated inventory files to pkg_name named
     const templateFiles = [
-      "inventory.yml",
-      "local-extras.yml",
-      "spatial-inventory.yml",
-      "spatial-local-extras.yml",
+      'inventory.yml',
+      'local-extras.yml',
+      'spatial-inventory.yml',
+      'spatial-local-extras.yml',
     ];
     for (var i = 0; i < templateFiles.length; i++) {
       const currentFile = `${dest}/quick-start-${templateFiles[i]}`;
@@ -985,12 +985,12 @@ module.exports = class extends Generator {
       const oldFile = `${dest}/${filePrefix}-inventory.yml`;
       if (debug) logger(`Trying to rename yml to ini: ${oldFile}`);
       if (this.fs.exists(oldFile)) {
-        var inventories = ["inventory", "local-extras", "local-passwords"];
+        var inventories = ['inventory', 'local-extras', 'local-passwords'];
         if (useSpatial) {
-          inventories.push("spatial-inventory", "spatial-local-extras");
+          inventories.push('spatial-inventory', 'spatial-local-extras');
         }
         if (useCAS) {
-          inventories.push("cas-inventory", "cas-local-extras");
+          inventories.push('cas-inventory', 'cas-local-extras');
         }
         let cmdResult;
         for (let i = 0; i < inventories.length; i++) {
@@ -1002,14 +1002,14 @@ module.exports = class extends Generator {
             const useGit = this.answers.LA_use_git;
             if (useGit) {
               cmdResult = this.spawnCommandSync(
-                "git",
-                ["mv", currentFile, destFile],
+                'git',
+                ['mv', currentFile, destFile],
                 cmdOpts
               );
               if (debug) logger(`git mv result ${JSON.stringify(cmdResult)}`);
             } else {
               cmdResult = this.spawnCommandSync(
-                "mv",
+                'mv',
                 [currentFile, destFile],
                 cmdOpts
               );
@@ -1028,10 +1028,10 @@ module.exports = class extends Generator {
 
       this.answers.LA_passwords = [];
       for (let num = 0; num < 50; num++) {
-        this.answers.LA_passwords.push(niceware.generatePassphrase(6).join(""));
+        this.answers.LA_passwords.push(niceware.generatePassphrase(6).join(''));
       }
 
-      const salt = bcrypt.genSaltSync(10, "a");
+      const salt = bcrypt.genSaltSync(10, 'a');
       this.answers.LA_admin_bcrypt_password = bcrypt.hashSync(
         this.answers.LA_passwords[23],
         salt
@@ -1044,7 +1044,7 @@ module.exports = class extends Generator {
       logger(
         `To create a different one modify this in: ${dest}/${filePrefix}-local-passwords.ini`
       );
-      logger("prior to deploy the CAS auth system.");
+      logger('prior to deploy the CAS auth system.');
     }
 
     this.answers.LA_apikeys = [];
@@ -1068,42 +1068,29 @@ module.exports = class extends Generator {
       this.answers
     );
 
-    if (useSpatial) {
-      if (!this.fs.exists(`${dest}/${filePrefix}-spatial-local-extras.ini`)) {
-        // When only create the extras inventory in the first run
-        this.fs.copyTpl(
-          this.templatePath(`quick-start-spatial-local-extras.ini`),
-          this.destinationPath(
-            `${dest}/${filePrefix}-spatial-local-extras.ini`
-          ),
-          this.answers
-        );
-      }
-      // .sample is always updated with new versions
-      this.fs.copyTpl(
-        this.templatePath(`quick-start-spatial-local-extras.ini`),
-        this.destinationPath(
-          `${dest}/${filePrefix}-spatial-local-extras.sample`
-        ),
-        this.answers
-      );
+    if (
+      useSpatial &&
+      this.fs.exists(`${dest}/${filePrefix}-spatial-inventory.ini`)
+    ) {
+      this.fs.delete(`${dest}/${filePrefix}-spatial-inventory.ini`);
     }
 
-    if (useCAS) {
-      if (!this.fs.exists(`${dest}/${filePrefix}-cas-local-extras.ini`)) {
-        // When only create the extras inventory in the first run
-        this.fs.copyTpl(
-          this.templatePath(`quick-start-cas-local-extras.ini`),
-          this.destinationPath(`${dest}/${filePrefix}-cas-local-extras.ini`),
-          this.answers
-        );
-      }
-      // .sample is always updated with new versions
-      this.fs.copyTpl(
-        this.templatePath(`quick-start-cas-local-extras.ini`),
-        this.destinationPath(`${dest}/${filePrefix}-cas-local-extras.sample`),
-        this.answers
-      );
+    if (
+      useSpatial &&
+      this.fs.exists(`${dest}/${filePrefix}-spatial-local-extras.sample`)
+    ) {
+      this.fs.delete(`${dest}/${filePrefix}-spatial-local-extras.sample`);
+    }
+
+    if (useCAS && this.fs.exists(`${dest}/${filePrefix}-cas-inventory.ini`)) {
+      this.fs.delete(`${dest}/${filePrefix}-cas-inventory.ini`);
+    }
+
+    if (
+      useCAS &&
+      this.fs.exists(`${dest}/${filePrefix}-cas-local-extras.sample`)
+    ) {
+      this.fs.delete(`${dest}/${filePrefix}-cas-local-extras.sample`);
     }
 
     this.fs.copyTpl(
@@ -1112,20 +1099,28 @@ module.exports = class extends Generator {
       this.answers
     );
 
-    if (useSpatial) {
-      this.fs.copyTpl(
-        this.templatePath("quick-start-spatial-inventory.ini"),
-        this.destinationPath(`${dest}/${filePrefix}-spatial-inventory.ini`),
-        this.answers
+    if (
+      this.fs.exists(`${dest}/${filePrefix}-spatial-local-extras.ini`) ||
+      this.fs.exists(`${dest}/${filePrefix}-cas-local-extras.ini`)
+    ) {
+      logger(
+        '-----------------------------------------------------------------------------------------------'
       );
-    }
-
-    if (useCAS) {
-      this.fs.copyTpl(
-        this.templatePath("quick-start-cas-inventory.ini"),
-        this.destinationPath(`${dest}/${filePrefix}-cas-inventory.ini`),
-        this.answers
+      logger(
+        `${em(
+          'WARNING'
+        )}: We are joining our inventories in a single one + a single local-extras.ini:`
       );
+      logger(
+        `Please join all your local-extras.ini files into a single ${dest}/${filePrefix}-local-extras.ini`
+      );
+      logger(
+        `Take into account that other old local-extras.ini will be ignored. Please delete them.`
+      );
+      logger(
+        '-----------------------------------------------------------------------------------------------'
+      );
+      logger('');
     }
 
     if (!this.fs.exists(`${dest}/${filePrefix}-local-passwords.ini`)) {
@@ -1142,7 +1137,7 @@ module.exports = class extends Generator {
     if (useBranding) {
       if (this.fs.exists(brandSettings)) {
         this.fs.copyTpl(
-          this.templatePath("base-branding-settings.js"),
+          this.templatePath('base-branding-settings.js'),
           this.destinationPath(`${brandSettings}.sample`),
           this.answers
         );
@@ -1152,11 +1147,11 @@ module.exports = class extends Generator {
         );
 
         let cmdResult = this.spawnCommandSync(
-          "git",
+          'git',
           [
-            "clone",
-            "--depth=1",
-            "https://github.com/living-atlases/base-branding.git",
+            'clone',
+            '--depth=1',
+            'https://github.com/living-atlases/base-branding.git',
             brandDest,
           ],
           {
@@ -1172,13 +1167,13 @@ module.exports = class extends Generator {
             `INFO: Do a "git submodule update --init --recursive --depth=1" in '${brandDest}' directory later`
           );
           // we remove the default branding settings
-          this.spawnCommandSync("rm", ["-f", brandSettings], {
+          this.spawnCommandSync('rm', ['-f', brandSettings], {
             cwd: this.destinationPath(),
             shell: true,
             stdio: null,
           });
           this.fs.copyTpl(
-            this.templatePath("base-branding-settings.js"),
+            this.templatePath('base-branding-settings.js'),
             this.destinationPath(brandSettings),
             this.answers
           );
@@ -1193,7 +1188,7 @@ module.exports = class extends Generator {
     }
 
     this.fs.copyTpl(
-      this.templatePath("ansiblew"),
+      this.templatePath('ansiblew'),
       this.destinationPath(`${dest}/ansiblew`),
       this.answers
     );
@@ -1215,23 +1210,23 @@ module.exports = class extends Generator {
     // this.installDependencies();
 
     if (firstRun && useGit) {
-      let cmdResult = this.spawnCommandSync("which", ["git"], cmdOpts);
+      let cmdResult = this.spawnCommandSync('which', ['git'], cmdOpts);
 
       if (cmdResult.status === 0) {
-        cmdResult = this.spawnCommandSync("git", ["status"], cmdOpts);
+        cmdResult = this.spawnCommandSync('git', ['status'], cmdOpts);
 
         if (cmdResult !== 0) {
-          cmdResult = this.spawnCommandSync("git", ["init"], cmdOpts);
-          cmdResult = this.spawnCommandSync("git", ["add", "--all"], cmdOpts);
+          cmdResult = this.spawnCommandSync('git', ['init'], cmdOpts);
+          cmdResult = this.spawnCommandSync('git', ['add', '--all'], cmdOpts);
           this.spawnCommandSync(
-            "git",
-            ["commit", "-am", '"Initial commit"'],
+            'git',
+            ['commit', '-am', '"Initial commit"'],
             cmdOpts
           );
         }
       } else {
         logger(
-          "Error: Please install git to track changes in your inventories"
+          'Error: Please install git to track changes in your inventories'
         );
       }
     }